requirements.txt
pyyaml
streamlit
pytest